IOS-4170 Move create account to the main screen
