+++ Anytype/Sources/Analytics/AnalyticsConstants.swift
@@ -195,7 +195,6 @@ enum ShowDeletionWarningRoute: String {
 }
 
 enum ScreenOnboardingStep: String {
-    case vault = "Vault"
     case phrase = "Phrase"
     case soul = "Soul"
 }
