+++ Anytype/Sources/PresentationLayer/Auth/Auth/AuthViewModelOutput.swift
@@ -2,7 +2,7 @@ import SwiftUI
 
 @MainActor
 protocol AuthViewModelOutput: AnyObject {
-    func onJoinAction() -> AnyView
+    func onJoinAction(state: JoinFlowState) -> AnyView
     func onLoginAction() -> AnyView
     func onSettingsAction() -> AnyView
 }
