+++ Anytype/Sources/PresentationLayer/Auth/JoinFlow/JoinFlowCoordinator.swift
@@ -2,7 +2,7 @@ import SwiftUI
 
 @MainActor
 protocol JoinFlowCoordinatorProtocol {
-    func startFlow() -> AnyView
+    func startFlow(state: JoinFlowState) -> AnyView
 }
 
 @MainActor
@@ -10,16 +10,14 @@ final class JoinFlowCoordinator: JoinFlowCoordinatorProtocol, JoinFlowOutput {
     
     // MARK: - JoinFlowCoordinatorProtocol
     
-    func startFlow() -> AnyView {
-        JoinFlowView(output: self).eraseToAnyView()
+    func startFlow(state: JoinFlowState) -> AnyView {
+        JoinFlowView(state: state, output: self).eraseToAnyView()
     }
     
     // MARK: - JoinFlowOutput
     
     func onStepChanged(_ step: JoinFlowStep, state: JoinFlowState, output: some JoinFlowStepOutput) -> AnyView {
         switch step {
-        case .vault:
-            return VaultView(state: state, output: output).eraseToAnyView()
         case .key:
             return KeyPhraseView(state: state, output: output).eraseToAnyView()
         case .soul:
